// libs/firefly-shared/src/state/document/user-profile/user-profile.actions.enum.ts
export enum ActionsUserProfile {
  Reset = '[UserProfile] Reset',
  Get = '[UserProfile] Get',
  Set = '[UserProfile] Set',
  Patch = '[UserProfile] Patch',
  PatchMetadata = '[UserProfile] Patch Metadata',
  Create = '[UserProfile] Create',
  Update = '[UserProfile] Update',
  Save = '[UserProfile] Save',
  Delete = '[UserProfile] Delete',
  SetId = '[UserProfile] SetId',
  Watch = '[UserProfile] Watch'
}